// src/app/core/Models/Interfaces/User/patient.model.ts
export interface RegisterPatientDto {
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  email: string;
  phoneNumber: string;
  gender: string;
  bloodGroup: string;
  address: string;
  city: string;
  stateId: number;
  countryId: number;
  zipCode: number;
  profileImage: string;
}